Move hardlink_or_copy to a common location so it can be reused.
